mark entire randombytes function as unsafe
Although randombytes, as currently implemented, is only meant to be
called from the tweetnacl C code, marking the entire function unsafe
ensures that calls from Rust code will be required to be wrapped in an
unsafe block. Semantically, this makes sense because the inherent
unsafety of the inputs makes any subsequent operations inside the
function unsafe and unable to be checked for safety.