do compile-time type checking of array lengths
Checking the length of arrays manually was pointless (for secret and
public keys) when this can just be done by adding a length parameter.
These lengths are constant and won't change, so we can just let the
compiler handle length checks for us.

This eliminates a couple possible errors, some boilerplate crap, and
a few unit tests.